requirements.txt
python-telegram-bot==20.7
PyPDF2==3.0.1
Pillow==9.5.0
python-pptx==0.6.23
fpdf==1.7.2
PyMuPDF==1.23.3
pdf2image==1.17.0
python-dotenv==1.0.1